kubernetes-operators/operator/requirements.txt
kubernetes>=24.2.0
PyYAML>=6.0